refactor: Seperating input.rs into submodules
