+++ src/input.rs
@@ -5,7 +5,7 @@ use crate::{
     double_click::{ClickCount, ClickState},
     events::CosmicTextChanged,
     prelude::*,
-    render_implementations::{RelativeQuery},
+    render_implementations::RelativeQuery,
 };
 use bevy::{
     ecs::{component::ComponentId, world::DeferredWorld},
@@ -16,33 +16,30 @@ use bevy::{
 };
 use cosmic_text::{Action, Cursor, Edit, Motion, Selection};
 
-#[cfg(target_arch = "wasm32")]
-use bevy::tasks::AsyncComputeTaskPool;
-#[cfg(target_arch = "wasm32")]
-#[allow(unused_imports)]
-use js_sys::Promise;
-#[cfg(target_arch = "wasm32")]
-use wasm_bindgen::prelude::*;
-#[cfg(target_arch = "wasm32")]
-use wasm_bindgen_futures::JsFuture;
+pub(crate) mod clipboard;
+pub(crate) mod keyboard;
+// mod hover;
+// mod click;
+// mod drag;
 
 /// System set for mouse and keyboard input events. Runs in [`PreUpdate`] and [`Update`]
 #[derive(SystemSet, Debug, Clone, PartialEq, Eq, Hash)]
-pub enum InputSet {
-    PreUpdate,
-    Update,
-}
+pub struct InputSet;
 
 pub(crate) struct InputPlugin;
 
 impl Plugin for InputPlugin {
     fn build(&self, app: &mut App) {
-        app.add_systems(PreUpdate, scroll.in_set(InputSet::PreUpdate))
+        app.add_systems(PreUpdate, scroll.in_set(InputSet))
             .add_systems(
                 Update,
-                (kb_move_cursor, kb_input_text, kb_clipboard)
+                (
+                    keyboard::kb_move_cursor,
+                    keyboard::kb_input_text,
+                    clipboard::kb_clipboard,
+                )
                     .chain()
-                    .in_set(InputSet::Update),
+                    .in_set(InputSet),
             );
 
         #[cfg(target_arch = "wasm32")]
@@ -52,22 +49,6 @@ impl Plugin for InputPlugin {
                 .add_systems(Update, poll_wasm_paste);
         }
     }
-}
-
-// TODO: hide this behind #cfg wasm, depends on wasm having own copy/paste fn
-/// Crossbeam channel struct for Wasm clipboard data
-#[cfg_attr(not(target_arch = "wasm32"), allow(dead_code))]
-pub(crate) struct WasmPaste {
-    text: String,
-    entity: Entity,
-}
-
-/// Async channel for receiving from the clipboard in Wasm
-#[cfg_attr(not(target_arch = "wasm32"), allow(dead_code))]
-#[derive(Resource)]
-pub(crate) struct WasmPasteAsyncChannel {
-    pub tx: crossbeam_channel::Sender<WasmPaste>,
-    pub rx: crossbeam_channel::Receiver<WasmPaste>,
 }
 
 #[derive(Component, Default, Debug)]
@@ -87,10 +68,10 @@ fn add_event_handlers(
     _component_id: ComponentId,
 ) {
     let mut observers = [
-        Observer::new(handle_click_sprite),
-        Observer::new(handle_dragstart_sprite),
-        Observer::new(handle_dragend_sprite),
-        Observer::new(handle_drag_sprite),
+        Observer::new(handle_click),
+        Observer::new(handle_dragstart),
+        Observer::new(handle_dragend),
+        Observer::new(handle_drag),
     ];
     for observer in &mut observers {
         observer.watch_entity(targeted_entity);
@@ -125,7 +106,7 @@ fn warn_no_editor_on_picking_event() {
     );
 }
 
-fn handle_click_sprite(
+fn handle_click(
     trigger: Trigger<Pointer<Click>>,
     mut editor: Query<(&mut InputState, &mut CosmicEditor, RelativeQuery)>,
     mut font_system: ResMut<CosmicFontSystem>,
@@ -202,7 +183,7 @@ fn handle_click_sprite(
     }
 }
 
-fn handle_dragstart_sprite(
+fn handle_dragstart(
     trigger: Trigger<Pointer<DragStart>>,
     mut editor: Query<(&mut InputState, &mut CosmicEditor, RelativeQuery)>,
     mut font_system: ResMut<CosmicFontSystem>,
@@ -246,7 +227,7 @@ fn handle_dragstart_sprite(
     }
 }
 
-fn handle_drag_sprite(
+fn handle_drag(
     trigger: Trigger<Pointer<Drag>>,
     mut editor: Query<(&InputState, &mut CosmicEditor)>,
     mut font_system: ResMut<CosmicFontSystem>,
@@ -285,7 +266,7 @@ fn handle_drag_sprite(
     }
 }
 
-fn handle_dragend_sprite(trigger: Trigger<Pointer<DragEnd>>, mut editor: Query<&mut InputState>) {
+fn handle_dragend(trigger: Trigger<Pointer<DragEnd>>, mut editor: Query<&mut InputState>) {
     let event = &trigger.event;
     let entity = trigger.target;
 
@@ -342,453 +323,5 @@ pub(crate) fn scroll(
                 }
             }
         }
-    }
-}
-
-pub(crate) fn kb_move_cursor(
-    active_editor: Res<FocusedWidget>,
-    keys: Res<ButtonInput<KeyCode>>,
-    mut cosmic_edit_query: Query<(&mut CosmicEditor,)>,
-    mut font_system: ResMut<CosmicFontSystem>,
-) {
-    let Some(active_editor_entity) = active_editor.0 else {
-        return;
-    };
-    if let Ok((mut editor,)) = cosmic_edit_query.get_mut(active_editor_entity) {
-        if keys.get_just_pressed().len() != 0 {
-            editor.cursor_visible = true;
-            editor.cursor_timer.reset();
-        }
-
-        let command = keypress_command(&keys);
-
-        #[cfg(target_arch = "wasm32")]
-        let command = if web_sys::window()
-            .unwrap()
-            .navigator()
-            .user_agent()
-            .unwrap_or("NoUA".into())
-            .contains("Macintosh")
-        {
-            keys.any_pressed([KeyCode::SuperLeft, KeyCode::SuperRight])
-        } else {
-            command
-        };
-
-        #[cfg(target_os = "macos")]
-        let option = keys.any_pressed([KeyCode::AltLeft, KeyCode::AltRight]);
-
-        let shift = keys.any_pressed([KeyCode::ShiftLeft, KeyCode::ShiftRight]);
-
-        // if shift key is pressed
-        let already_has_selection = editor.selection() != Selection::None;
-        if shift && !already_has_selection {
-            let cursor = editor.cursor();
-            editor.set_selection(Selection::Normal(cursor));
-        }
-
-        #[cfg(target_os = "macos")]
-        let should_jump = command && option;
-        #[cfg(not(target_os = "macos"))]
-        let should_jump = command;
-
-        if should_jump && keys.just_pressed(KeyCode::ArrowLeft) {
-            editor.action(&mut font_system.0, Action::Motion(Motion::PreviousWord));
-            if !shift {
-                editor.set_selection(Selection::None);
-            }
-            return;
-        }
-        if should_jump && keys.just_pressed(KeyCode::ArrowRight) {
-            editor.action(&mut font_system.0, Action::Motion(Motion::NextWord));
-            if !shift {
-                editor.set_selection(Selection::None);
-            }
-            return;
-        }
-        if should_jump && keys.just_pressed(KeyCode::Home) {
-            editor.action(&mut font_system.0, Action::Motion(Motion::BufferStart));
-            if !shift {
-                editor.set_selection(Selection::None);
-            }
-            return;
-        }
-        if should_jump && keys.just_pressed(KeyCode::End) {
-            editor.action(&mut font_system.0, Action::Motion(Motion::BufferEnd));
-            if !shift {
-                editor.set_selection(Selection::None);
-            }
-            return;
-        }
-
-        if keys.just_pressed(KeyCode::ArrowLeft) {
-            editor.action(&mut font_system.0, Action::Motion(Motion::Left));
-            if !shift {
-                editor.set_selection(Selection::None);
-            }
-            return;
-        }
-        if keys.just_pressed(KeyCode::ArrowRight) {
-            editor.action(&mut font_system.0, Action::Motion(Motion::Right));
-            if !shift {
-                editor.set_selection(Selection::None);
-            }
-            return;
-        }
-        if keys.just_pressed(KeyCode::ArrowUp) {
-            editor.action(&mut font_system.0, Action::Motion(Motion::Up));
-            if !shift {
-                editor.set_selection(Selection::None);
-            }
-            return;
-        }
-        if keys.just_pressed(KeyCode::ArrowDown) {
-            editor.action(&mut font_system.0, Action::Motion(Motion::Down));
-            if !shift {
-                editor.set_selection(Selection::None);
-            }
-            return;
-        }
-        if keys.just_pressed(KeyCode::Escape) {
-            editor.action(&mut font_system.0, Action::Escape);
-        }
-        if command && keys.just_pressed(KeyCode::KeyA) {
-            editor.action(&mut font_system.0, Action::Motion(Motion::BufferEnd));
-            let current_cursor = editor.cursor();
-            editor.set_selection(Selection::Normal(Cursor {
-                line: 0,
-                index: 0,
-                affinity: current_cursor.affinity,
-            }));
-            return;
-        }
-        if keys.just_pressed(KeyCode::Home) {
-            editor.action(&mut font_system.0, Action::Motion(Motion::Home));
-            if !shift {
-                editor.set_selection(Selection::None);
-            }
-            return;
-        }
-        if keys.just_pressed(KeyCode::End) {
-            editor.action(&mut font_system.0, Action::Motion(Motion::End));
-            if !shift {
-                editor.set_selection(Selection::None);
-            }
-            return;
-        }
-        if keys.just_pressed(KeyCode::PageUp) {
-            editor.action(&mut font_system.0, Action::Motion(Motion::PageUp));
-            if !shift {
-                editor.set_selection(Selection::None);
-            }
-            return;
-        }
-        if keys.just_pressed(KeyCode::PageDown) {
-            editor.action(&mut font_system.0, Action::Motion(Motion::PageDown));
-            if !shift {
-                editor.set_selection(Selection::None);
-            }
-        }
-    }
-}
-
-pub(crate) fn kb_input_text(
-    active_editor: Res<FocusedWidget>,
-    keys: Res<ButtonInput<KeyCode>>,
-    mut char_evr: EventReader<KeyboardInput>,
-    mut cosmic_edit_query: Query<(
-        &mut CosmicEditor,
-        &mut CosmicEditBuffer,
-        &MaxLines,
-        &MaxChars,
-        Entity,
-        Option<&ReadOnly>,
-    )>,
-    mut evw_changed: EventWriter<CosmicTextChanged>,
-    mut font_system: ResMut<CosmicFontSystem>,
-    mut is_deleting: Local<bool>,
-) {
-    let Some(active_editor_entity) = active_editor.0 else {
-        return;
-    };
-
-    if let Ok((mut editor, buffer, max_lines, max_chars, entity, readonly_opt)) =
-        cosmic_edit_query.get_mut(active_editor_entity)
-    {
-        let command = keypress_command(&keys);
-        if keys.get_just_pressed().len() != 0 {
-            editor.cursor_visible = true;
-            editor.cursor_timer.reset();
-        }
-        let readonly = readonly_opt.is_some();
-
-        if keys.just_pressed(KeyCode::Backspace) & !readonly {
-            // fix for issue #8
-            let select = editor.selection();
-            match select {
-                Selection::Line(cursor) => {
-                    if editor.cursor().line == cursor.line && editor.cursor().index == cursor.index
-                    {
-                        editor.set_selection(Selection::None);
-                    }
-                }
-                Selection::Normal(cursor) => {
-                    if editor.cursor().line == cursor.line && editor.cursor().index == cursor.index
-                    {
-                        editor.set_selection(Selection::None);
-                    }
-                }
-                Selection::Word(cursor) => {
-                    if editor.cursor().line == cursor.line && editor.cursor().index == cursor.index
-                    {
-                        editor.set_selection(Selection::None);
-                    }
-                }
-                Selection::None => {}
-            }
-
-            *is_deleting = true;
-        }
-
-        if keys.just_released(KeyCode::Backspace) {
-            *is_deleting = false;
-        }
-        if keys.just_pressed(KeyCode::Delete) && !readonly {
-            editor.action(&mut font_system.0, Action::Delete);
-            editor.with_buffer_mut(|b| b.set_redraw(true));
-        }
-
-        if readonly {
-            return;
-        }
-
-        let mut is_edit = false;
-        let mut is_return = false;
-        if keys.just_pressed(KeyCode::Enter) {
-            is_return = true;
-            if (max_lines.0 == 0 || buffer.lines.len() < max_lines.0)
-                && (max_chars.0 == 0 || buffer.get_text().len() < max_chars.0)
-            {
-                // to have new line on wasm rather than E
-                is_edit = true;
-                editor.action(&mut font_system.0, Action::Insert('\n'));
-            }
-        }
-
-        if !is_return {
-            for char_ev in char_evr.read() {
-                is_edit = true;
-                if *is_deleting {
-                    editor.action(&mut font_system.0, Action::Backspace);
-                } else if !command
-                    && (max_chars.0 == 0 || buffer.get_text().len() < max_chars.0)
-                    && matches!(char_ev.state, bevy::input::ButtonState::Pressed)
-                {
-                    match &char_ev.logical_key {
-                        Key::Character(char) => {
-                            let b = char.as_bytes();
-                            for c in b {
-                                let c: char = (*c).into();
-                                editor.action(&mut font_system.0, Action::Insert(c));
-                            }
-                        }
-                        Key::Space => {
-                            editor.action(&mut font_system.0, Action::Insert(' '));
-                        }
-                        _ => (),
-                    }
-                }
-            }
-        }
-
-        if !is_edit {
-            return;
-        }
-
-        evw_changed.send(CosmicTextChanged((
-            entity,
-            editor.with_buffer_mut(|b| b.get_text()),
-        )));
-    }
-}
-
-pub(crate) fn kb_clipboard(
-    active_editor: Res<FocusedWidget>,
-    keys: Res<ButtonInput<KeyCode>>,
-    mut evw_changed: EventWriter<CosmicTextChanged>,
-    mut font_system: ResMut<CosmicFontSystem>,
-    mut cosmic_edit_query: Query<(
-        &mut CosmicEditor,
-        &mut CosmicEditBuffer,
-        &MaxLines,
-        &MaxChars,
-        Entity,
-        Option<&ReadOnly>,
-    )>,
-    _channel: Option<Res<WasmPasteAsyncChannel>>,
-) {
-    let Some(active_editor_entity) = active_editor.0 else {
-        return;
-    };
-
-    if let Ok((mut editor, buffer, max_lines, max_chars, entity, readonly_opt)) =
-        cosmic_edit_query.get_mut(active_editor_entity)
-    {
-        let command = keypress_command(&keys);
-
-        let readonly = readonly_opt.is_some();
-
-        let mut is_clipboard = false;
-        #[cfg(not(target_arch = "wasm32"))]
-        {
-            if let Ok(mut clipboard) = arboard::Clipboard::new() {
-                if command && keys.just_pressed(KeyCode::KeyC) {
-                    if let Some(text) = editor.copy_selection() {
-                        clipboard.set_text(text).unwrap();
-                        return;
-                    }
-                }
-                if command && keys.just_pressed(KeyCode::KeyX) && !readonly {
-                    if let Some(text) = editor.copy_selection() {
-                        clipboard.set_text(text).unwrap();
-                        editor.delete_selection();
-                    }
-                    is_clipboard = true;
-                }
-                if command && keys.just_pressed(KeyCode::KeyV) && !readonly {
-                    if let Ok(text) = clipboard.get_text() {
-                        for c in text.chars() {
-                            if max_chars.0 == 0 || buffer.get_text().len() < max_chars.0 {
-                                if c == 0xA as char {
-                                    if max_lines.0 == 0 || buffer.lines.len() < max_lines.0 {
-                                        editor.action(&mut font_system.0, Action::Insert(c));
-                                    }
-                                } else {
-                                    editor.action(&mut font_system.0, Action::Insert(c));
-                                }
-                            }
-                        }
-                    }
-                    is_clipboard = true;
-                }
-            }
-        }
-
-        #[cfg(target_arch = "wasm32")]
-        {
-            if command && keys.just_pressed(KeyCode::KeyC) {
-                if let Some(text) = editor.copy_selection() {
-                    write_clipboard_wasm(text.as_str());
-                    return;
-                }
-            }
-
-            if command && keys.just_pressed(KeyCode::KeyX) && !readonly {
-                if let Some(text) = editor.copy_selection() {
-                    write_clipboard_wasm(text.as_str());
-                    editor.delete_selection();
-                }
-                is_clipboard = true;
-            }
-            if command && keys.just_pressed(KeyCode::KeyV) && !readonly {
-                let tx = _channel.unwrap().tx.clone();
-                let _task = AsyncComputeTaskPool::get().spawn(async move {
-                    let promise = read_clipboard_wasm();
-
-                    let result = JsFuture::from(promise).await;
-
-                    if let Ok(js_text) = result {
-                        if let Some(text) = js_text.as_string() {
-                            let _ = tx.try_send(WasmPaste { text, entity });
-                        }
-                    }
-                });
-
-                return;
-            }
-        }
-
-        if !is_clipboard {
-            return;
-        }
-
-        evw_changed.send(CosmicTextChanged((entity, buffer.get_text())));
-    }
-}
-
-fn keypress_command(keys: &ButtonInput<KeyCode>) -> bool {
-    #[cfg(target_os = "macos")]
-    let command = keys.any_pressed([KeyCode::SuperLeft, KeyCode::SuperRight]);
-
-    #[cfg(not(target_os = "macos"))]
-    let command = keys.any_pressed([KeyCode::ControlLeft, KeyCode::ControlRight]);
-
-    #[cfg(target_arch = "wasm32")]
-    let command = if web_sys::window()
-        .unwrap()
-        .navigator()
-        .user_agent()
-        .unwrap_or("NoUA".into())
-        .contains("Macintosh")
-    {
-        keys.any_pressed([KeyCode::SuperLeft, KeyCode::SuperRight])
-    } else {
-        command
-    };
-
-    command
-}
-
-#[cfg(target_arch = "wasm32")]
-#[wasm_bindgen]
-pub fn write_clipboard_wasm(text: &str) {
-    let clipboard = web_sys::window().unwrap().navigator().clipboard();
-    let _result = clipboard.write_text(text);
-}
-
-#[cfg(target_arch = "wasm32")]
-#[wasm_bindgen]
-pub fn read_clipboard_wasm() -> Promise {
-    let clipboard = web_sys::window().unwrap().navigator().clipboard();
-    clipboard.read_text()
-}
-
-#[cfg(target_arch = "wasm32")]
-pub(crate) fn poll_wasm_paste(
-    channel: Res<WasmPasteAsyncChannel>,
-    mut editor_q: Query<
-        (
-            &mut CosmicEditor,
-            &mut CosmicEditBuffer,
-            &MaxChars,
-            &MaxChars,
-        ),
-        Without<ReadOnly>,
-    >,
-    mut evw_changed: EventWriter<CosmicTextChanged>,
-    mut font_system: ResMut<CosmicFontSystem>,
-) {
-    let inlet = channel.rx.try_recv();
-    match inlet {
-        Ok(inlet) => {
-            let entity = inlet.entity;
-            if let Ok((mut editor, buffer, max_chars, max_lines)) = editor_q.get_mut(entity) {
-                let text = inlet.text;
-                for c in text.chars() {
-                    if max_chars.0 == 0 || buffer.get_text().len() < max_chars.0 {
-                        if c == 0xA as char {
-                            if max_lines.0 == 0 || buffer.lines.len() < max_lines.0 {
-                                editor.action(&mut font_system.0, Action::Insert(c));
-                            }
-                        } else {
-                            editor.action(&mut font_system.0, Action::Insert(c));
-                        }
-                    }
-                }
-
-                evw_changed.send(CosmicTextChanged((entity, buffer.get_text())));
-            }
-        }
-        Err(_) => {}
     }
 }
