+++ src/render_implementations.rs
@@ -1,10 +1,13 @@
 //! Generalizes over render target implementations.
 //!
-//! ## Sprite:
+//! All implementations should use [`bevy::picking`] for interactions,
+//! even [`SourceType::Ui`], for consistency.
+//!
+//! ## Sprite: [`TextEdit2d`]
 //! Requires [`Sprite`] component and requires [`Sprite.custom_size`] to be Some( non-zero )
 //!
-//! ## UI:
-//! Requires [`ImageNode`] for rendering and [`Button`] for [`Interaction`]s
+//! ## UI: [`TextEdit`]
+//! Requires [`ImageNode`] for rendering
 // TODO: Remove `CosmicWidgetSize`?
 
 mod prelude {
@@ -152,57 +155,4 @@ pub(crate) fn get_node_cursor_pos(
                 }
             }),
     })
-}
-
-pub(crate) fn hover_sprites(
-    windows: Query<&Window, With<bevy::window::PrimaryWindow>>,
-    mut cosmic_edit_query: Query<
-        (&mut Sprite, &Visibility, &GlobalTransform, &HoverCursor),
-        With<CosmicEditBuffer>,
-    >,
-    camera_q: Query<(&Camera, &GlobalTransform), CameraFilter>,
-    mut hovered: Local<bool>,
-    mut last_hovered: Local<bool>,
-    mut evw_hover_in: EventWriter<TextHoverIn>,
-    mut evw_hover_out: EventWriter<TextHoverOut>,
-) {
-    *hovered = false;
-    if windows.iter().len() == 0 {
-        return;
-    }
-    let window = windows.single();
-    let (camera, camera_transform) = camera_q.single();
-
-    let mut icon = CursorIcon::System(SystemCursorIcon::Default);
-
-    for (sprite, visibility, node_transform, hover) in &mut cosmic_edit_query.iter_mut() {
-        if visibility == Visibility::Hidden {
-            continue;
-        }
-
-        let size = sprite.custom_size.unwrap_or(Vec2::ONE);
-        if crate::render_implementations::get_node_cursor_pos(
-            window,
-            node_transform,
-            size,
-            SourceType::Sprite,
-            camera,
-            camera_transform,
-        )
-        .is_some()
-        {
-            *hovered = true;
-            icon = hover.0.clone();
-        }
-    }
-
-    if *last_hovered != *hovered {
-        if *hovered {
-            evw_hover_in.send(TextHoverIn(icon));
-        } else {
-            evw_hover_out.send(TextHoverOut);
-        }
-    }
-
-    *last_hovered = *hovered;
 }
